# docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Enclii Platform API
  description: |
    Open source DevOps platform API for deploying and managing containerized applications.

    ## Overview

    Enclii Switchyard is the control plane API that manages projects, services, deployments,
    domains, teams, and infrastructure. It provides a comprehensive REST API with 107+ endpoints
    covering the complete application lifecycle.

    ## Authentication

    Most endpoints require JWT authentication. There are two authentication modes:

    - **OIDC Mode** (Production): Uses Janua SSO with RS256 JWT validation
    - **Local Mode** (Development): Uses self-signed JWT with password-based auth

    Include the token in the `Authorization` header as `Bearer <token>`.

    ## Rate Limiting

    - Auth endpoints: 10 requests/minute (5 for register/login)
    - Standard endpoints: 100 requests/minute
    - Build endpoints: 20 requests/minute

    ## Environments

    Services can be deployed to multiple environments:
    - `dev` - Development (default)
    - `staging` - Pre-production testing
    - `prod` - Production

    ## WebSocket Endpoints

    Real-time log streaming is available via WebSocket:
    - `ws://api.enclii.dev/v1/services/{id}/logs/stream` - Service logs
    - `ws://api.enclii.dev/v1/deployments/{id}/logs/stream` - Deployment logs
    - `ws://api.enclii.dev/v1/services/{id}/builds/{build_id}/logs/stream` - Build logs

  version: 1.0.0
  contact:
    name: MADFAM Engineering
...
servers:
...
paths:
...
  /projects/{slug}:
    get:
      summary: Get project
      description: Get project details by slug.
      tags: [projects]
      operationId: getProject
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: my-project
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found

  # ============================================
  # CI RUNNER CONFIGURATION
  # ============================================
...
  /projects/{slug}/services:
    get:
      summary: List services
      description: Get all services in a project.
      tags: [services]
      operationId: listServices
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
    post:
      summary: Create service
      description: Create a new service in a project.
      tags: [services]
      operationId: createService
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /projects/{slug}/services/bulk:
    post:
      summary: Bulk create services
      description: |
        Create multiple services at once. Useful for monorepos where multiple
        services share the same git repository.
      tags: [services]
      operationId: bulkCreateServices
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateServicesRequest'
      responses:
        '201':
          description: Services created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCreateServicesResponse'
  /services/{id}:
    get:
      summary: Get service
      description: Get service details by ID.
      tags: [services]
      operationId: getService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          description: Service not found
    patch:
      summary: Update service
      description: Update service configuration.
      tags: [services]
      operationId: updateService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: Service updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    $ref: '#/components/schemas/Service'
                  message:
                    type: string
    delete:
      summary: Delete service
      description: Delete a service and all its resources. Requires admin role.
      tags: [services]
      operationId: deleteService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or OIDC callback
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        git_repo:
          type: string
          description: GitHub HTTPS URL
        git_branch:
          type: string
        app_path:
          type: string
          description: Path in monorepo
        auto_deploy:
          type: boolean
        auto_deploy_branch:
          type: string
        auto_deploy_env:
          type: string
        build_config:
          $ref: '#/components/schemas/BuildConfig'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP response headers injected via nginx ingress annotations
          example:
            Cross-Origin-Opener-Policy: same-origin
            Cross-Origin-Embedder-Policy: require-corp
        status:
          type: string
          enum: [running, stopped, building]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BuildConfig:
      type: object
      properties:
        type:
          type: string
          enum: [auto, dockerfile, buildpack]
        dockerfile:
          type: string
        buildpack:
          type: string
    CreateServiceRequest:
      type: object
      required:
        - name
        - git_repo
      properties:
        name:
          type: string
        git_repo:
          type: string
        app_path:
          type: string
        auto_deploy:
          type: boolean
          default: true
        auto_deploy_branch:
          type: string
          default: main
        build_config:
          $ref: '#/components/schemas/BuildConfig'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP response headers injected via nginx ingress annotations
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        git_repo:
          type: string
        app_path:
          type: string
        auto_deploy:
          type: boolean
        auto_deploy_branch:
          type: string
        auto_deploy_env:
          type: string
        build_config:
          $ref: '#/components/schemas/BuildConfig'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP response headers injected via nginx ingress annotations
    BulkCreateServicesRequest:
      type: object
      required:
        - git_repo
        - services
      properties:
        git_repo:
          type: string
        git_branch:
          type: string
          default: main
        services:
          type: array
          items:
            type: object
            required:
              - name
              - app_path
            properties:
              name:
                type: string
              app_path:
                type: string
              port:
                type: integer
              build_command:
                type: string
              start_command:
                type: string
              auto_deploy:
                type: boolean
                default: true
              auto_deploy_branch:
                type: string
              auto_deploy_env:
                type: string
    BulkCreateServicesResponse:
      type: object
      properties:
        created:
          type: integer
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
...
security:
  - bearerAuth: []